+++ spec/samples/reversed_fornum.lua
@@ -1,0 +1,3 @@
+for i = #(...), -1.5 do
+   print(i)
+end